+++ contract-deployer/src/app/layout.tsx
@@ -1,14 +1,8 @@
-"use client";
-
-import { Inter } from "next/font/google";
+import { AppComponent } from "@/components/App";
 import "./globals.css";
-import { MintbaseWalletContextProvider } from "@mintbase-js/react";
-import { MintbaseWalletSetup } from "@/config/setup";
+
 import "@near-wallet-selector/modal-ui/styles.css";
 import { Metadata } from "next";
-
-const inter = Inter({ subsets: ["latin"] });
-
 
 export const metadata: Metadata = {
   title: `Near Contract Deployer`,
@@ -40,14 +34,6 @@ export default function RootLayout({
   children: React.ReactNode;
 }) {
   return (
-    <MintbaseWalletContextProvider {...MintbaseWalletSetup}>
-      <html lang="en">
-        <body className={`${inter.className} dark`}>
-          <div className="flex flex-1 flex-col min-h-screen text-gray-500 gradient w-full  h-full flex justify-center items-center bold text-white">
-            {children}
-          </div>
-        </body>
-      </html>
-    </MintbaseWalletContextProvider>
+    <AppComponent> {children} </AppComponent>
   );
 }
